Update libbpf.c to be compiler agnostic
1) Some compilers cannot do "+" on a void* since they don't know how
   large an entry is, since "+ n" would be interpreted as literally
   adding "n * sizeof(void)".  Instead, cast to "char*" before addition.

2) "= {}" is not legal in some compilers, need to do "= {0}" to
   initialize a structure to be zeroed.

3) Directly casting a pointer to an integer results in a compiler
   warning on some compilers, but one can cast it to intptr_t or
   uintptr_t and then cast to an integer type.

4) Wrap GCC-specific pragmas in ifdef __GNUC__

5) ":?" is also not legal in some compilers, need a token in between

+++ include/linux/err.h
@@ -8,26 +8,26 @@
 
 #define MAX_ERRNO       4095
 
-#define IS_ERR_VALUE(x) ((x) >= (unsigned long)-MAX_ERRNO)
+#define IS_ERR_VALUE(x) ((x) >= (uintptr_t)-MAX_ERRNO)
 
 static inline void * ERR_PTR(long error_)
 {
-	return (void *) error_;
+	return (void *) (intptr_t) error_;
 }
 
 static inline long PTR_ERR(const void *ptr)
 {
-	return (long) ptr;
+	return (long) (intptr_t) ptr;
 }
 
 static inline bool IS_ERR(const void *ptr)
 {
-	return IS_ERR_VALUE((unsigned long)ptr);
+	return IS_ERR_VALUE((uintptr_t)ptr);
 }
 
 static inline bool IS_ERR_OR_NULL(const void *ptr)
 {
-	return (!ptr) || IS_ERR_VALUE((unsigned long)ptr);
+	return (!ptr) || IS_ERR_VALUE((uintptr_t)ptr);
 }
 
 static inline long PTR_ERR_OR_ZERO(const void *ptr)
